+++ src/main/java/com/midokura/midolman/mgmt/rest_api/v1/resources/BgpResource.java
@@ -20,16 +20,22 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriInfo;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.midokura.midolman.mgmt.auth.AuthManager;
+import com.midokura.midolman.mgmt.auth.UnauthorizedException;
+import com.midokura.midolman.mgmt.data.OwnerQueryable;
 import com.midokura.midolman.mgmt.data.dao.BgpZkManagerProxy;
+import com.midokura.midolman.mgmt.data.dao.PortZkManagerProxy;
 import com.midokura.midolman.mgmt.data.dto.Bgp;
 import com.midokura.midolman.mgmt.rest_api.v1.resources.AdRouteResource.BgpAdRouteResource;
 import com.midokura.midolman.state.Directory;
 import com.midokura.midolman.state.StateAccessException;
+import com.midokura.midolman.state.ZkStateSerializationException;
 
 /**
  * Root resource class for bgps.
@@ -62,14 +68,25 @@ public class BgpResource extends RestResource {
      *            BGP UUID.
      * @return Bgp object.
      * @throws StateAccessException
+     * @throws UnauthorizedException
+     * @throws ZkStateSerializationException
      */
     @GET
     @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Bgp get(@PathParam("id") UUID id) throws StateAccessException {
+    public Bgp get(@PathParam("id") UUID id, @Context SecurityContext context)
+            throws StateAccessException, ZkStateSerializationException,
+            UnauthorizedException {
+
         // Get a bgp for the given ID.
         BgpZkManagerProxy dao = new BgpZkManagerProxy(zooKeeper, zookeeperRoot,
                 zookeeperMgmtRoot);
+
+        if (!AuthManager.isOwner(context, dao, id)) {
+            throw new UnauthorizedException(
+                    "Can only see your own advertised route.");
+        }
+
         Bgp bgp = null;
         try {
             bgp = dao.get(id);
@@ -86,15 +103,19 @@ public class BgpResource extends RestResource {
     @PUT
     @Path("{id}")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response update(@PathParam("id") UUID id, Bgp bgp)
-            throws StateAccessException {
+    public Response update(@PathParam("id") UUID id, Bgp bgp,
+            @Context SecurityContext context) throws StateAccessException,
+            ZkStateSerializationException, UnauthorizedException {
         BgpZkManagerProxy dao = new BgpZkManagerProxy(zooKeeper, zookeeperRoot,
                 zookeeperMgmtRoot);
+
+        if (!AuthManager.isOwner(context, dao, id)) {
+            throw new UnauthorizedException(
+                    "Can only see your own advertised route.");
+        }
+
         try {
             dao.update(id, bgp);
-        } catch (StateAccessException e) {
-            log.error("Error accessing data", e);
-            throw e;
         } catch (Exception e) {
             log.error("Unhandled error", e);
             throw new UnknownRestApiException(e);
@@ -104,9 +125,17 @@ public class BgpResource extends RestResource {
 
     @DELETE
     @Path("{id}")
-    public void delete(@PathParam("id") UUID id) throws StateAccessException {
+    public void delete(@PathParam("id") UUID id,
+            @Context SecurityContext context) throws StateAccessException,
+            ZkStateSerializationException, UnauthorizedException {
         BgpZkManagerProxy dao = new BgpZkManagerProxy(zooKeeper, zookeeperRoot,
                 zookeeperMgmtRoot);
+
+        if (!AuthManager.isOwner(context, dao, id)) {
+            throw new UnauthorizedException(
+                    "Can only see your own advertised route.");
+        }
+
         try {
             dao.delete(id);
         } catch (StateAccessException e) {
@@ -141,15 +170,30 @@ public class BgpResource extends RestResource {
             this.zookeeperMgmtRoot = zkMgmtRootDir;
         }
 
+        private boolean isPortOwner(SecurityContext context)
+                throws StateAccessException, ZkStateSerializationException {
+            OwnerQueryable q = new PortZkManagerProxy(zooKeeper, zookeeperRoot,
+                    zookeeperMgmtRoot);
+            return AuthManager.isOwner(context, q, portId);
+        }
+
         /**
          * Index of bgps belonging to the port.
          * 
          * @return A list of bgps.
          * @throws StateAccessException
+         * @throws UnauthorizedException
+         * @throws ZkStateSerializationException
          */
         @GET
         @Produces(MediaType.APPLICATION_JSON)
-        public List<Bgp> list() throws StateAccessException {
+        public List<Bgp> list(@Context SecurityContext context)
+                throws StateAccessException, ZkStateSerializationException,
+                UnauthorizedException {
+            if (!isPortOwner(context)) {
+                throw new UnauthorizedException("Can only see your own BGP.");
+            }
+
             BgpZkManagerProxy dao = new BgpZkManagerProxy(zooKeeper,
                     zookeeperRoot, zookeeperMgmtRoot);
             List<Bgp> bgps = null;
@@ -171,15 +215,22 @@ public class BgpResource extends RestResource {
          * @param bgp
          *            Bgp object mapped to the request input.
          * @throws StateAccessException
+         * @throws UnauthorizedException
+         * @throws ZkStateSerializationException
          * @returns Response object with 201 status code set if successful.
          */
         @POST
         @Consumes(MediaType.APPLICATION_JSON)
-        public Response create(Bgp bgp, @Context UriInfo uriInfo)
-                throws StateAccessException {
-            bgp.setPortId(portId);
+        public Response create(Bgp bgp, @Context UriInfo uriInfo,
+                @Context SecurityContext context) throws StateAccessException,
+                ZkStateSerializationException, UnauthorizedException {
+            if (!isPortOwner(context)) {
+                throw new UnauthorizedException("Can only create your own BGP.");
+            }
             BgpZkManagerProxy dao = new BgpZkManagerProxy(zooKeeper,
                     zookeeperRoot, zookeeperMgmtRoot);
+            bgp.setPortId(portId);
+
             UUID id = null;
             try {
                 id = dao.create(bgp);
